refactor(examples): decompose AfterBurner backdrop into per-layer Renderables
`SkyboxStage` used to override `draw()` and inline every backdrop
visual (sky / ground gradient, two mountain ranges, ground bands)
inside that single override. Two problems:

1. **Architecture.** None of those visuals were Renderables in
   their own right — they couldn't be re-used, individually
   toggled, or sorted via the normal world tree.

2. **Perf.** The bigger one. The sky + ground gradients were built
   via `renderer.createLinearGradient` every frame, which inside
   the engine rasterizes a power-of-two canvas via Canvas2D's
   `createLinearGradient` + `fillStyle` + `fillRect`. For a 1024×768
   viewport that's ~2 M gradient-baked pixels per frame purely for
   colors that never change — ~120 M/sec at 60 fps, and the bulk
   of the previously-reported ~9 ms stage draw cost.

New `backdrop/` directory:

- `SkyGradient.ts` — sky + ground gradients baked once into a
  `1 × 1024` canvas each at construction; per-frame draw is two
  `drawImage` blits stretched vertically to the current horizon.
  Color resolution stays high because the bake resolution is
  generous; horizontal stretch from 1 px to (w + 2*pad) is a single
  solid color so no aliasing.
- `MountainHorizon.ts` — owns the already-baked mountain layers
  (moved out of SkyboxStage). Same six `drawImage` blits per frame
  as before, just self-contained now.
- `GroundGrid.ts` — owns `scrollT` + `scrollPaused`; gates the
  scroll advance on `state.isPaused()` like before.
- `horizonMath.ts` — shared `computeHorizonY(camera, h)` so all
  three layers land on the same pitch-driven horizon pixel.
- `BackdropContainer.ts` — floating Container that holds the three
  layers. Floating wrap means the engine sets the screen ortho
  projection ONCE per frame; the camera roll rotation is also
  applied once at the container level, then `super.draw()` walks
  the children with that transform already active. Depth = 10000
  so painter-sort paints us first.

`SkyboxStage.ts` is now ~25 lines: spawn a BackdropContainer in
`onResetEvent`, forward `setScrollPaused` to the grid.
`GameController` updated to call `stage.setScrollPaused(true/false)`
instead of reaching into the stage's field directly.
